requirement.txt
Flask-Session-0.3.1
redis-3.3.8
WTForms-2.2.1 flask-wtf-0.14.2 # 使用里面的CSRF的防护机制
flask-migrate
Pillow